+++ src/Api/Handler.hs
@@ -7,11 +7,9 @@ module Api.Handler (
 
 import Control.Monad.Catch (throwM)
 import Control.Monad.Logger (logInfoNS)
-import Data.Map qualified as Map
 import Data.String.ToString (toString)
 import Data.Text (Text)
 import Data.Text qualified as Text
-import Data.Vector qualified as Vector
 import Network.WebSockets qualified as WebSockets
 import Servant (err404, err500)
 import Servant.API.BasicAuth (BasicAuthData (BasicAuthData))
@@ -32,7 +30,6 @@ import Api.Error (JsonError (JsonError), throwJsonError)
 import Api.Types (
   ControlApiAuthData (ControlApiAuthData),
   GetDatumByHashResponse (GetDatumByHashResponse),
-  GetDatumsByHashesDatum (GetDatumsByHashesDatum),
   GetDatumsByHashesRequest (GetDatumsByHashesRequest),
   GetDatumsByHashesResponse (GetDatumsByHashesResponse),
   SetDatumFilterRequest (SetDatumFilterRequest),
@@ -97,10 +94,8 @@ datumServiceHandlers =
       App GetDatumsByHashesResponse
     getDatumsByHashes (GetDatumsByHashesRequest hashes) = do
       datums <- Database.getDatumsByHashes hashes >>= catchDatabaseError
-      let (_, rightDatums) = Map.mapEither id datums
       pure $
-        GetDatumsByHashesResponse
-          (uncurry GetDatumsByHashesDatum <$> (Vector.fromList . Map.toList) rightDatums)
+        GetDatumsByHashesResponse datums
 
     getTx :: Text -> App Text
     getTx txId = do